tools/portable/requirements.txt
# Minimal deps for download-only mode (run_download.py)
curl_cffi>=0.7
beautifulsoup4>=4.12
lxml>=5.0
pyyaml>=6.0
